sre-agent/runtime/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
openai>=1.0.0

# Google auth — ADC token management for Vertex AI LLM provider
google-auth>=2.29.0

# OpenTelemetry — metrics, traces, and logs export via OTLP
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
